src/cst/models/requirements.txt
"""
Setup script and requirements for CST implementation
"""

# requirements.txt content
REQUIREMENTS = """
# Core ML/DL dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
torch-audio>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
datasets>=2.12.0
tokenizers>=0.13.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
psutil>=5.9.0
redis>=4.5.0
requests>=2.31.0
Pillow>=9.5.0

# Development tools
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0

# Optional dependencies for specific features
# faiss-gpu>=1.7.4  # For efficient similarity search
# opencv-python>=4.7.0  # For image processing
# librosa>=0.10.0  # For audio processing
# sentence-transformers>=2.2.0  # For text embeddings
"""